added usage search for transfer stations

+++ sophena/src/sophena/rcp/editors/basedata/transfer/stations/TransferStationEditor.java
@@ -14,6 +14,8 @@ import org.eclipse.ui.forms.widgets.ScrolledForm;
 import org.eclipse.ui.forms.widgets.Section;
 
 import sophena.db.daos.RootEntityDao;
+import sophena.db.usage.SearchResult;
+import sophena.db.usage.UsageSearch;
 import sophena.model.ProductType;
 import sophena.model.TransferStation;
 import sophena.rcp.App;
@@ -21,6 +23,7 @@ import sophena.rcp.Icon;
 import sophena.rcp.M;
 import sophena.rcp.editors.Editor;
 import sophena.rcp.editors.basedata.BaseTableLabel;
+import sophena.rcp.editors.basedata.UsageError;
 import sophena.rcp.utils.Actions;
 import sophena.rcp.utils.Editors;
 import sophena.rcp.utils.KeyEditorInput;
@@ -132,12 +135,11 @@ public class TransferStationEditor extends Editor {
 					"Soll die ausgewählte Hausübergabestation wirklich gelöscht werden?");
 			if (!doIt)
 				return;
-			// TODO: usage search
-			// List<SearchResult> usage = new UsageSearch(App.getDb()).of(s);
-			// if (!usage.isEmpty()) {
-			// UsageError.show(usage);
-			// return;
-			// }
+			List<SearchResult> usage = new UsageSearch(App.getDb()).of(s);
+			if (!usage.isEmpty()) {
+				UsageError.show(usage);
+				return;
+			}
 			try {
 				dao.delete(s);
 				stations.remove(s);
